Add v2 dashboard pages and remove legacy v1 routes
Adds new top-level routes (commitments, ministries, policy-areas, faq, feed, platform),
TrackerNav component, and updated homepage. Removes old [department] and v2/ routes.
