--- app/[department]/page.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-export default function Page() {
-  return <></>;
-}
