+++ app/policy-areas/[slug]/page.tsx
@@ -43,7 +43,7 @@ export default async function PolicyAreaPage({
       {/* Breadcrumb + Header */}
       <div>
         <Link
-          href="/v2"
+          href="/"
           className="text-xs text-gray-400 hover:text-[#8b2332] transition-colors"
         >
           &larr; Overview
@@ -72,7 +72,7 @@ export default async function PolicyAreaPage({
             {commitments.map((c: CommitmentListing) => (
               <Link
                 key={c.id}
-                href={`/v2/commitments/${c.id}`}
+                href={`/commitments/${c.id}`}
                 className="block bg-white border border-[#cdc4bd] hover:border-gray-400 transition-colors p-4"
               >
                 <div className="flex items-start gap-3">
